+++ project/Build.scala
@@ -18,28 +18,12 @@ import sbt._
 import Keys._
 import net.liftweb.sbt.LiftBuildPlugin._
 import Dependencies._
+import LiftSbtHelpers._
 
 import com.typesafe.sbt.web.SbtWeb
 import com.typesafe.sbt.web.SbtWeb.autoImport._
 
-
 object BuildDef extends Build {
-
-  /**
-   * A helper that returns the revision and JAR file for a given dependency.
-   * Useful when trying to attach API doc URI information.
-   */
-  def findManagedDependency(classpath: Seq[Attributed[File]],
-                            organization: String,
-                            name: String): Option[(String,File)] = {
-    classpath.collectFirst {
-      case entry @ MatchingModule(moduleOrganization, moduleName, revision)
-          if moduleOrganization == organization &&
-             moduleName.startsWith(name) =>
-        (revision, entry.data)
-    }
-  }
-
   lazy val liftProjects = core ++ web ++ persistence
 
   lazy val framework =
@@ -224,37 +208,4 @@ object BuildDef extends Build {
         .settings(parallelExecution in Test := false)
 
 
-  def coreProject = liftProject("core") _
-  def webProject = liftProject("web") _
-  def persistenceProject = liftProject("persistence") _
-
-  /** Project definition helper that simplifies creation of `ProjectReference`.
-    *
-    * It is a convenience method to create a Lift `ProjectReference` module by having the boilerplate for most common
-    * activities tucked in.
-    *
-    * @param base     the base path location of project module.
-    * @param prefix   the prefix of project module.
-    * @param module   the name of the project module. Typically, a project id is of the form lift-`module`.
-    */
-  def liftProject(base: String, prefix: String = "lift-")(module: String): Project =
-    liftProject(id = if (module.startsWith(prefix)) module else prefix + module,
-                base = file(base) / module.stripPrefix(prefix))
-
-  def liftProject(id: String, base: File): Project = {
-    Project(id, base)
-      .settings(liftBuildSettings: _*)
-      .settings(scalacOptions ++= List("-feature", "-language:implicitConversions"))
-      .settings(
-        autoAPIMappings := true,
-        apiMappings ++= {
-          val cp: Seq[Attributed[File]] = (fullClasspath in Compile).value
-
-          findManagedDependency(cp, "org.scala-lang.modules", "scala-xml").map {
-            case (revision, file)  =>
-              (file -> url("http://www.scala-lang.org/api/" + version))
-          }.toMap
-        }
-      )
-  }
 }
